Made the forked repo PSR compliant and composer installable.
