+++ phpunit.xml.dist
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<phpunit colors="true" bootstrap="vendor/autoload.php">
+	<testsuites>
+		<testsuite name="Oath Testsuite">
+			<directory>test</directory>
+		</testsuite>
+	</testsuites>
+</phpunit>
